Removed references to tracking admin pages
Due to changes in the way that Drupal 7.x displays tracking pages to the user it is not currently possible to track them without a complete overhaul and the use of a feature of the LiveStats tracking code which is not yet officially supported.

This is something which could be implemented at a later date if there is any requests for it.