Re-implemented a different style fix for #1123760
This fix now uses the setting part of drupal_add_js() to add a custom setting which is therefore not vulnerable to XSS attacks.

Also found and fixed a bug whereby guest users will see an error because of an undefined variable in some cases. This has now been fixed by declaring a boolean to say whether that variable is set and whether or not to add the user code to the tracking code.

Signed-off-by: Aaron Parker <parkeyparker@gmail.com>